Added connector and refactored
