+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.bereavementsupportpaymentapi.controllers.MicroserviceHelloWorldController.hello
+GET        /nino-info             uk.gov.hmrc.bereavementsupportpaymentapi.controllers.ServiceController.getNinoInfo
