+++ test/uk/gov/hmrc/bereavementsupportpaymentapi/controllers/ServiceControllerSpec.scala
@@ -6,14 +6,14 @@ import play.api.http.Status
 import play.api.test.Helpers._
 import play.api.test.{FakeRequest, Helpers}
 
-class MicroserviceHelloWorldControllerSpec extends AnyWordSpec with Matchers {
+class ServiceControllerSpec extends AnyWordSpec with Matchers {
 
   private val fakeRequest = FakeRequest("GET", "/")
-  private val controller = new MicroserviceHelloWorldController(Helpers.stubControllerComponents())
+  private val controller = new ServiceController(Helpers.stubControllerComponents())
 
   "GET /" should {
     "return 200" in {
-      val result = controller.hello()(fakeRequest)
+      val result = controller.getNinoInfo()(fakeRequest)
       status(result) shouldBe Status.OK
     }
   }
